requirements.txt
python-dateutil
requests
PyYAML
